Virtual class for Woodbury identity matrices

The WoodburyMatrix is a virtual class, contained by both GWoodburyMatrix (for general matrices) and SWoodburyMatrix (for symmetric matrices). See WoodburyMatrix for construction of these classes. The methods available for these classes are described below; see also the solve methods. This class is itself a subclass of Matrix, so basic matrix methods like nrow, ncol, dim and so on also work.

# S4 method for GWoodburyMatrix,ANY
%*%(x, y)

# S4 method for SWoodburyMatrix,ANY
%*%(x, y)

# S4 method for GWoodburyMatrix
t(x)

# S4 method for SWoodburyMatrix
t(x)

Arguments

x
WoodburyMatrix object

y
Matrix or vector

Functions

GWoodburyMatrix-class: Sub-class representing a generic matrix.

SWoodburyMatrix-class: Sub-class representing a symmetric matrix.

%*%,GWoodburyMatrix,ANY-method: Matrix multiplication (generally fast and

%*%,SWoodburyMatrix,ANY-method: Matrix multiplication (generally fast and

t,GWoodburyMatrix-method: Return the transpose of the matrix as another GWoodburyMatrix.

t,SWoodburyMatrix-method: Does nothing, just returns x.

Slots

A

n x n subclass of Matrix (GWoodburyMatrix) or symmetricMatrix (SWoodburyMatrix).

B

p x p subclass of Matrix (GWoodburyMatrix) or symmetricMatrix (SWoodburyMatrix).

U

n x p subclass of Matrix (only for

V

p x m subclass of Matrix (only for

X

n x p subclass of Matrix (only for

O

p x p subclass of Matrix

See also